# app/OpenApi/Schemas/TagSchema.php
<?php

namespace App\OpenApi\Schemas;

use OpenApi\Annotations as OA;

/**
 * @OA\Schema(
 *     schema="Tag",
 *     title="Tag Model",
 *     description="Representa uma tag associada a posts",
 *     @OA\Property(property="id", type="integer", readOnly=true, example=1),
 *     @OA\Property(property="name", type="string", example="Laravel"),
 *     @OA\Property(property="slug", type="string", readOnly=true, example="laravel"),
 *     @OA\Property(property="created_at", type="string", format="date-time", readOnly=true),
 *     @OA\Property(property="updated_at", type="string", format="date-time", readOnly=true)
 * )
 */
class TagSchema {}